+++ controllers/userController.js
@@ -1,21 +1,15 @@
 /* eslint-disable linebreak-style */
-const { log } = require("debug");
 const users = require("../models/users");
 const {
   hashPassword,
   isPasswordValid,
   tokengen
 } = require("../helpers/authHelper");
-const Student = require("../models/Student");
-const Admin = require("../models/admin");
-const Parent = require("../models/parent");
-const Resource = require("../models/resource");
+const checkRole = require("../helpers/roleHelper");
 
-const { hashPassword, isPasswordValid, tokengen } = require("../helpers/authHelper");
 const { generateMailForSignup } = require("../services/email/mailhelper");
 const mailingService = require("../services/email/mailingservice");
 const { Roles } = require("../helpers/constants");
-
 
 exports.signUp = async (req, res) => {
   try {
@@ -45,48 +39,20 @@ exports.signUp = async (req, res) => {
     const hash = await hashPassword(password);
 
     // save the user details
-
     const createUser = users({
       email,
       firstname,
       lastname,
-      role,
+      role: finishedrole,
       password: hash
     });
     // Check user role andpopulate different collection
-    const checkRole = async (Users) => {
-      const uniqueId = Users._id;
-      if (Users.role === "Student") {
-        const student = await Student({ studentId: uniqueId });
-        student.save();
-      } else if (Users.role === "Resource") {
-        const resource = await Resource({ resourceId: uniqueId });
-        resource.save();
-      } else if (Users.role === "Parent") {
-        const parent = await Parent({ parentId: uniqueId });
-        parent.save();
-      } else if (Users.role === "Admin") {
-        const admin = await Admin({ adminId: uniqueId });
-        admin.save();
-      }
-    };
     checkRole(createUser);
 
     // Save User to Database
-    createUser.save();
-    // TODO
-    // send a welcome maile to the user
-    return res
-      .status(201)
-      .json({ response: "user credentials succesfully saved", createUser });
-
-
-    const createUser = await users.create({
-      email, firstname, lastname, role: finishedrole, password: hash
-    });
+    await createUser.save();
 
     const loginLink = "https://excelmind.com/users/login";
-    // TODO
     // send a welcome mail to the user
     const options = {
       receiver: email,
@@ -95,9 +61,9 @@ exports.signUp = async (req, res) => {
       output: generateMailForSignup(loginLink, email)
     };
     await mailingService(options);
-
-    return res.status(201).json({ response: "user credentials succesfully saved", data: createUser });
-
+    return res
+      .status(201)
+      .json({ response: "user credentials succesfully saved", data: createUser });
   } catch (error) {
     return res.status(500).json({ response: `error ${error} occured` });
   }
@@ -116,8 +82,6 @@ exports.login = async (req, res) => {
     if (!checkPassword) {
       return res.status(403).json({ response: "wrong password" });
     }
-
-
 
     // eslint-disable-next-line no-underscore-dangle
     const token = await tokengen({ userId: user._id });
